fix: correct cache key (#54)
This package uses a cache to check if a tag exists or if it needs
to be created. Once a tag is created it is added to the cache. This
did not work properly, since the different keys were used at cache
lookup (tag name) and add to cache (fully qualified tag name).